airflow/requirements.txt
# Airflow and core dependencies
apache-airflow[spark,hdfs]==2.8.1
pyspark==3.5.0

# Data processing libraries
requests>=2.31.0
hdfs3>=0.3.1

# Additional useful libraries for data processing
pandas>=2.0.0
numpy>=1.24.0

# Logging and monitoring
python-json-logger>=2.0.0

# Development and testing (optional)
pytest>=7.0.0
flake8>=6.0.0 